Split vtscore/training/thresholds.py into a five-module package
The module held five non-overlapping concerns in 2680 lines. A reference
graph over its top-level symbols shows the layering is already a clean DAG
with zero back edges, so the split is a move, not a redesign:

  knobs -> gmm -> {anchored, conformal} -> blend

* knobs      - what an Inclusion value, a Train/Calibrate split and the
               sentinels mean.  Imports nothing.
* gmm        - the 1-D two-component mixture, its anchored variant, and the
               cut rules.  Imports no sibling at all.
* anchored   - the shipped path: per-fold anchored mixtures combined in
               quantile space, plus the voted-media haystack exclusion.
* conformal  - the rival estimator: fold splits and split-conformal
               quantiles over pooled held-out scores.
* blend      - the retired safe-threshold blend, kept as the anchored
               path's small-label fallback.

The package __init__ re-exports every previously importable top-level name
(private ones included), so vtscore.training.thresholds.X resolves exactly
as before and no caller, doc reference or lazy import changed.

Two follow-ons the move forced:

* calculate_gmm_threshold and fit_gmm_threshold carried byte-identical
  cut logic.  Both now live in gmm.py and the former delegates to the
  latter, so they cannot drift.  Behaviour is unchanged, including the
  0.5 and median fallbacks.
* Rebinding a name on the package no longer reaches the submodule that
  calls it, so the six tests that stub an internal now patch the
  submodule that uses it.  Reads off the package still resolve; only
  setattr targets moved.

check-eval-app-sync.py resolves py: mirrors by parsing the file at the
dotted path, not by importing, so its three thresholds mirrors name the
submodule now.  The digests are unchanged (the moves are verbatim), so
the pins file needed no --update.

tests_lib/sorting/test_thresholds_package.py pins the three invariants
that must survive: every submodule top-level name is re-exported and
identical to the submodule attribute, the sibling-import graph has no
back edges, and calculate_gmm_threshold matches fit_gmm_threshold
bit-for-bit on the fallback paths as well as the fitted one.

Refs #3381
